Move get-locale to lib/geo - duplicate of country-code
